API/Portfolio/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
gunicorn==20.0.4
Pillow==8.1.0
pip==21.0.1
psycopg2==2.8.6
pytz==2021.1
setuptools==53.0.0
sqlparse==0.4.1
wheel==0.36.2
